Altered Workshop Client, Entity and Factory
