+++ src/main/java/za/ac/cput/frontendcarservicemanagement/factory/WorkshopFactory.java
@@ -1,5 +1,9 @@
 package za.ac.cput.frontendcarservicemanagement.factory;
 
+/*Author
+  Moegamad Nur Duncan 220014442
+* */
+
 import za.ac.cput.frontendcarservicemanagement.entity.Workshop;
 import za.ac.cput.frontendcarservicemanagement.util.Helper;
 
@@ -13,5 +17,9 @@ public class WorkshopFactory {
                 .name(name)
                 .location(location)
                 .build();
+    }
+
+    public static Workshop createWorkshop(long workshopID, String name, String location) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 }
